double version of chi_square added

+++ core/vnl/algo/vnl_chi_squared.h
@@ -16,7 +16,7 @@
 
 
 //: Compute cumulative distribution function value for chi-squared distribution
-extern float vnl_chi_squared_cumulative(float chisq, int dof);
+extern double vnl_chi_squared_cumulative(double chisq, int dof);
 
 //------------------------------------------------------------
